fix(api): emit lock-expiry side effects from the sweep
Mirror release-desktop-edit-lock: after expiring a session, push a
session-closed SSE event, close its desktop SSE connections, and broadcast
an ["entities", workspaceId] invalidation per workspace. Without this,
clients with the entities view or a desktop stream open keep showing the
stale lock until an unrelated refetch. Extract buildExpiryAuditEvents into
a pure module so the unit test does not import the eager Redis client in
lib/sse.
