+++ apps/api/src/lib/scheduler/tasks/desktop-edit-session-expiry.ts
@@ -3,56 +3,19 @@ import { and, asc, eq, inArray, lt } from "drizzle-orm";
 
 import { rootDb } from "@/api/db/root";
 import { auditLogs, desktopEditSessions, workspaces } from "@/api/db/schema";
-import { AUDIT_ACTION, AUDIT_RESOURCE_TYPE } from "@/api/lib/audit-log";
-import type { SafeId } from "@/api/lib/branded-types";
+import {
+  closeSessionConnections,
+  pushSessionEvent,
+} from "@/api/handlers/entities/desktop-edit-session-events";
+import { buildExpiryAuditEvents } from "@/api/lib/scheduler/tasks/desktop-edit-session-expiry-audit";
 import type { SchedulerTask } from "@/api/lib/scheduler/types";
+import { broadcast } from "@/api/lib/sse";
 
 export const EXPIRE_DESKTOP_EDIT_SESSIONS_TASK =
   "desktopEditSessions.expire" as const;
 
 /** Sweep in bounded batches so each transaction stays small. */
 const EXPIRE_SWEEP_BATCH_SIZE = 200;
-
-type ExpirableSession = {
-  id: string;
-  workspaceId: SafeId<"workspace">;
-  organizationId: SafeId<"organization">;
-  createdBy: string;
-};
-
-type ExpiryAuditEvent = {
-  organizationId: SafeId<"organization">;
-  workspaceId: SafeId<"workspace">;
-  userId: string;
-  action: typeof AUDIT_ACTION.UPDATE;
-  resourceType: typeof AUDIT_RESOURCE_TYPE.DESKTOP_EDIT_SESSION;
-  resourceId: string;
-  changes: { status: { old: "open"; new: "expired" } };
-  metadata: { reason: "token_expired" };
-};
-
-/**
- * A row selected as expirable can drop out before the UPDATE commits: a
- * concurrent finalize or takeover leaves "open", or a checkpoint/resume
- * renews tokenExpiresAt. The ids the UPDATE actually transitioned are the
- * source of truth for what to audit, not the originally selected batch.
- */
-export const buildExpiryAuditEvents = (
-  sessions: readonly ExpirableSession[],
-  expiredIds: ReadonlySet<string>,
-): ExpiryAuditEvent[] =>
-  sessions
-    .filter((session) => expiredIds.has(session.id))
-    .map((session) => ({
-      organizationId: session.organizationId,
-      workspaceId: session.workspaceId,
-      userId: session.createdBy,
-      action: AUDIT_ACTION.UPDATE,
-      resourceType: AUDIT_RESOURCE_TYPE.DESKTOP_EDIT_SESSION,
-      resourceId: session.id,
-      changes: { status: { old: "open", new: "expired" } },
-      metadata: { reason: "token_expired" },
-    }));
 
 /**
  * Transition abandoned desktop edit sessions whose token TTL has lapsed
@@ -97,7 +60,7 @@ export const expireDesktopEditSessions: SchedulerTask = async ({
 
     const batchIds = batch.map((session) => session.id);
 
-    const sweptCount = await rootDb.transaction(async (tx) => {
+    const expiredSessions = await rootDb.transaction(async (tx) => {
       const transitioned = await tx
         .update(desktopEditSessions)
         .set({ status: "expired", closedAt: now })
@@ -120,10 +83,30 @@ export const expireDesktopEditSessions: SchedulerTask = async ({
         await tx.insert(auditLogs).values(auditEvents);
       }
 
-      return transitioned.length;
+      return batch.filter((session) => expiredIds.has(session.id));
     });
 
-    expired += sweptCount;
+    // Post-commit, cross-instance side effects, mirroring
+    // release-desktop-edit-lock: tell desktop SSE streams the session is
+    // gone and nudge open entities views to refetch, so the lock clears
+    // for active clients instead of lingering until an unrelated refetch.
+    for (const session of expiredSessions) {
+      pushSessionEvent(session.id, {
+        type: "session-closed",
+        data: { reason: "expired" },
+      });
+      closeSessionConnections(session.id);
+    }
+    for (const workspaceId of new Set(
+      expiredSessions.map((session) => session.workspaceId),
+    )) {
+      broadcast(workspaceId, {
+        type: "invalidate-query",
+        data: ["entities", workspaceId],
+      });
+    }
+
+    expired += expiredSessions.length;
 
     if (batch.length < EXPIRE_SWEEP_BATCH_SIZE) {
       break;
